adjusts specs post-fork so they pass
There were a few places that changes around the initial fork broke the
specs -- things like the repo being in a new place, the executable
having moved, and the version having been bumped. I think it makes
sense to update the specs to match those changes, since I understand
the intention is to maintain all project practices except the name.